+++ src/sections/privacyPolicy.js
@@ -1,26 +1,27 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui";
 import { Container, Box, Heading, Text, Image, Button } from "theme-ui";
-import BannerImg from "assets/banner-thumb.png";
 import ShapeLeft from "assets/shape-left.png";
 import ShapeRight from "assets/shape-right.png";
+import TextFeature from 'components/text-feature';
+
+const data = {
+	subTitle: 'Core features',
+	title: 'Smart Jackpots that you may love this anytime & anywhere',
+	description:
+	    `Get your tests delivered at let home collect sample from the victory of
+        the managements that supplies best design system guidelines ever.`,
+};
 
 export default function PrivacyPolicy() {
 	return (
-		<section sx={styles.banner} id="home">
+		<section sx={styles.banner} id="policy">
 			<Container sx={styles.banner.container}>
 				<Box sx={styles.banner.contentBox}>
-					<Heading as="h1" variant="heroPrimary">
-						Redefining the Athletic Community
-					</Heading>
-					<Text as="p" variant="heroSecondary">
-						Coming soon...
-					</Text>
-					{/* <Button variant="primary">Explore</Button> */}
-				</Box>
-
-				<Box sx={styles.banner.imageBox}>
-					<Image src={BannerImg} alt="banner" />
+                    <TextFeature
+                        title={data.title}
+                        description={data.description}
+                    />
 				</Box>
 			</Container>
 		</section>
